pr_6/requirements.txt
psycopg2-binary==2.9.9
hvac==2.3.0
